Move key mapping to its own cpp/h
