+++ vs/FPSci.lib.vcxproj
@@ -145,6 +145,7 @@
   <ItemGroup>
     <ClInclude Include="..\source\FPSciApp.h" />
     <ClInclude Include="..\source\CombatText.h" />
+    <ClInclude Include="..\source\KeyMapping.h" />
     <ClInclude Include="..\source\Session.h" />
     <ClInclude Include="..\source\ConfigFiles.h" />
     <ClInclude Include="..\source\Logger.h" />
@@ -161,6 +162,7 @@
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="..\source\FPSciApp.cpp" />
+    <ClCompile Include="..\source\KeyMapping.cpp" />
     <ClCompile Include="..\source\Session.cpp" />
     <ClCompile Include="..\source\Logger.cpp" />
     <ClCompile Include="..\source\PhysicsScene.cpp" />
