+++ vs/FPSci.lib.vcxproj.filters
@@ -63,6 +63,9 @@
     <ClInclude Include="..\source\StartupConfig.h">
       <Filter>Header Files</Filter>
     </ClInclude>
+    <ClInclude Include="..\source\KeyMapping.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="..\source\sqlHelpers.cpp">
@@ -96,6 +99,9 @@
       <Filter>Source Files</Filter>
     </ClCompile>
     <ClCompile Include="..\source\StartupConfig.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="..\source\KeyMapping.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
   </ItemGroup>
